# backend/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem:
    properties:
      code:
        type: string
      detail:
        type: string
      meta:
        additionalProperties: {}
        type: object
      request_id:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_application.EpisodeMapping:
    properties:
      episode_id:
        type: string
      inherit_status:
        type: string
      previous_episode_id:
        type: string
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.EpisodeText:
    properties:
      content_hash:
        type: string
      episode_id:
        type: string
      script_version_id:
        type: string
      span_end:
        type: integer
      span_start:
        type: integer
      text:
        type: string
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.EpisodeView:
    properties:
      candidate:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.SplitSet'
      episodes:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.Episode'
        type: array
      head:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.VersionHead'
      version_id:
        type: string
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceHistoryPage:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceSummary'
        type: array
      next_revision:
        type: integer
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceSnapshotDetail:
    properties:
      char_count:
        type: integer
      content_hash:
        type: string
      document:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.RichDocument'
      id:
        type: string
      media_asset_id:
        type: string
      origin:
        type: string
      original_html:
        type: string
      plain_text:
        type: string
      position:
        type: integer
      previous_source_id:
        type: string
      provenance:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.SourceProvenance'
      rich_sha256:
        type: string
      source_kind:
        type: string
      source_lineage_id:
        type: string
      source_revision:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_application.SplitReceipt:
    properties:
      confirmation_id:
        type: string
      episode_mappings:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.EpisodeMapping'
        type: array
      episodes:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.Episode'
        type: array
      project_revision:
        type: integer
      renamed_episode_ids:
        items:
          type: string
        type: array
      script_revision:
        type: integer
      split_revision:
        type: integer
      split_set_id:
        type: string
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.StructurePage:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.StructureSummary'
        type: array
      next_version_no:
        type: integer
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.StructureReceipt:
    properties:
      episode_revision:
        type: integer
      project_revision:
        type: integer
      review_status:
        type: string
      script_revision:
        type: integer
      structure_id:
        type: string
      version_no:
        type: integer
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.StructureSummary:
    properties:
      created_at:
        type: string
      episode_id:
        type: string
      id:
        type: string
      source_hash:
        type: string
      version_no:
        type: integer
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.VersionHead:
    properties:
      candidate_split_set_id:
        type: string
      confirmed_split_set_id:
        type: string
      split_revision:
        type: integer
      version_id:
        type: string
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.VersionPage:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.VersionSummary'
        type: array
      next_version_no:
        type: integer
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_application.VersionSummary:
    properties:
      char_count:
        type: integer
      content_hash:
        type: string
      created_at:
        type: string
      document_sha256:
        type: string
      id:
        type: string
      source_count:
        type: integer
      source_manifest_sha256:
        type: string
      version_no:
        type: integer
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.Episode:
    properties:
      confirmed_structure_id:
        type: string
      current_structure_id:
        type: string
      id:
        type: string
      inherit_status:
        type: string
      is_delete:
        type: boolean
      org_id:
        type: string
      previous_episode_id:
        type: string
      project_id:
        type: string
      revision:
        type: integer
      script_version_id:
        type: string
      seq_no:
        type: integer
      span_end:
        type: integer
      span_start:
        type: integer
      split_set_id:
        type: string
      title:
        type: string
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.EpisodeBoundary:
    properties:
      seq_no:
        type: integer
      source_lineage_id:
        type: string
      span_end:
        type: integer
      span_start:
        type: integer
      title:
        type: string
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.EpisodeStructure:
    properties:
      actor_id:
        type: string
      created_at:
        type: string
      document:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.StructureDocument'
      episode_id:
        type: string
      id:
        type: string
      org_id:
        type: string
      project_id:
        type: string
      source_hash:
        type: string
      version_no:
        type: integer
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.RichDocument:
    properties:
      attrs:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.RichAttrs'
      content:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.RichDocument'
        type: array
      marks:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.RichMark'
        type: array
      text:
        type: string
      type:
        type: string
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.ScalarSpan:
    properties:
      end:
        type: integer
      start:
        type: integer
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.SourceProvenance:
    properties:
      encoding:
        type: string
      external_id:
        type: string
      external_position:
        type: integer
      mapping:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.SourceOriginMapping'
        type: array
      warnings:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.SourceExtractionWarning'
        type: array
    type: object
...
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.SplitSet:
    properties:
      boundaries:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.EpisodeBoundary'
        type: array
      created_at:
        type: string
      id:
        type: string
      kind:
        type: string
      org_id:
        type: string
      origin:
        type: string
      preface:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.ScalarSpan'
      project_id:
        type: string
      version_id:
        type: string
      warnings:
        items:
          type: string
        type: array
    type: object
  github_com_StephenQiu30_lanverse_backend_internal_script_domain.StructureDocument:
    properties:
      scenes:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.StructureScene'
        type: array
      unassigned_lines:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.UnassignedLine'
        type: array
    type: object
...
  internal_script_adapter_http.SplitReviewRequest:
    properties:
      ack_invalidate:
        type: boolean
      boundaries:
        items:
          $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.EpisodeBoundary'
        type: array
      candidate_set_id:
        type: string
      expected_revision:
        type: integer
      expected_split_revision:
        type: integer
      preface:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.ScalarSpan'
      version_id:
        type: string
    type: object
  internal_script_adapter_http.StructureConfirmRequest:
    properties:
      ack_invalidate:
        type: boolean
      base_structure_version_no:
        type: integer
      expected_episode_revision:
        type: integer
      expected_revision:
        type: integer
    type: object
  internal_script_adapter_http.StructureDetail:
    properties:
      episode:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.Episode'
      structure:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.EpisodeStructure'
    type: object
  internal_script_adapter_http.StructureSaveRequest:
    properties:
      base_structure_version_no:
        type: integer
      document:
        $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_domain.StructureDocument'
      expected_episode_revision:
        type: integer
      expected_revision:
        type: integer
    type: object
...
info:
  contact: {}
  description: Single workspace project query, resource canvas commands and authorized
    media previews.
  title: Lanverse API
  version: "0.1"
paths:
...
  /api/episodes/{eid}/source-text:
    get:
      operationId: getScriptEpisodeSourceText
      parameters:
      - description: 正式分集UUID
        in: path
        name: eid
        required: true
        type: string
      - description: 包含起点，Unicode scalar
        in: query
        name: from
        required: true
        type: integer
      - description: 不包含终点，最多65536scalar
        in: query
        name: to
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.EpisodeText'
      summary: 读取分集规范正文区间
      tags:
      - script
  /api/episodes/{eid}/structure:
    get:
      operationId: getScriptEpisodeStructure
      parameters:
      - description: 正式分集UUID
        in: path
        name: eid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_script_adapter_http.StructureDetail'
      summary: 当前手工结构
      tags:
      - script
    post:
      consumes:
      - application/json
      operationId: saveScriptEpisodeStructure
      parameters:
      - description: 正式分集UUID
        in: path
        name: eid
        required: true
        type: string
      - description: UUID
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: 完整结构与两级CAS
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_script_adapter_http.StructureSaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.StructureReceipt'
      summary: 保存完整手工结构候选
      tags:
      - script
  /api/episodes/{eid}/structure/confirm:
    post:
      consumes:
      - application/json
      operationId: confirmScriptEpisodeStructure
      parameters:
      - description: 正式分集UUID
        in: path
        name: eid
        required: true
        type: string
      - description: UUID
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: 当前结构CAS与影响确认
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_script_adapter_http.StructureConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.StructureReceipt'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
      summary: 明确确认当前手工结构
      tags:
      - script
  /api/episodes/{eid}/structure/versions:
    get:
      operationId: listScriptEpisodeStructureVersions
      parameters:
      - description: 正式分集UUID
        in: path
        name: eid
        required: true
        type: string
      - description: 严格早于结构版本号
        in: query
        name: before_version_no
        type: integer
      - description: 1..100
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.StructurePage'
      summary: 手工结构全部版本
      tags:
      - script
  /api/episodes/{eid}/structure/versions/{version}:
    get:
      operationId: getScriptEpisodeStructureVersion
      parameters:
      - description: 正式分集UUID
        in: path
        name: eid
        required: true
        type: string
      - description: 结构版本号
        in: path
        name: version
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_script_adapter_http.StructureDetail'
      summary: 指定历史手工结构
      tags:
      - script
...
  /api/projects/{pid}/episodes:
    get:
      operationId: listScriptEpisodes
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 剧本版本UUID
        in: query
        name: version_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.EpisodeView'
      summary: 指定版本的候选与正式分集
      tags:
      - script
  /api/projects/{pid}/episodes/split/confirm:
    post:
      consumes:
      - application/json
      operationId: confirmScriptEpisodeSplit
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: UUID
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: 完整分集边界及影响确认
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_script_adapter_http.SplitReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.SplitReceipt'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_platform_httpapi.Problem'
      summary: 确认完整分集边界与身份映射
      tags:
      - script
  /api/projects/{pid}/episodes/split/resplit:
    post:
      consumes:
      - application/json
      operationId: resplitScriptEpisodes
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: UUID
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: 当前头与候选修订
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_script_adapter_http.SplitReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.SplitReceipt'
      summary: 按正文标题规则重新生成候选分集
      tags:
      - script
...
  /api/projects/{pid}/script-source-snapshots/{sid}:
    get:
      operationId: getScriptSourceSnapshot
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 来源快照UUID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceSnapshotDetail'
      summary: 读取指定不可变章节快照
      tags:
      - script
...
  /api/projects/{pid}/script-sources/{lineage}/history:
    get:
      operationId: listScriptSourceHistory
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 稳定来源UUID
        in: path
        name: lineage
        required: true
        type: string
      - description: 严格早于来源修订
        in: query
        name: before_revision
        type: integer
      - description: 1..100
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.SourceHistoryPage'
      summary: 稳定章节身份的全部快照
      tags:
      - script
...
  /api/projects/{pid}/script-versions:
    get:
      operationId: listScriptVersions
      parameters:
      - description: 项目UUID
        in: path
        name: pid
        required: true
        type: string
      - description: 严格早于版本号
        in: query
        name: before_version_no
        type: integer
      - description: 1..100
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_StephenQiu30_lanverse_backend_internal_script_application.VersionPage'
      summary: 完整剧本版本历史
      tags:
      - script
...
swagger: "2.0"